% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{alex}
\alias{alex}
\title{Catch Insensitive, Inconsiderate Writing}
\usage{
alex(value = NULL)
}
\arguments{
\item{value}{If it is a connection object, then we read all
lines from it with \code{readLines()} and then run \code{alex}
on them. Otherwise if it is a character vector, we run \code{alex}
on it. If \code{NULL}, then we run \code{alex} on all markdown
(.Rmd and .md) files in the current working directory.}
}
\value{
An object with S3 class \code{alex}. It's \code{messages}
  entry contains the suggestions for changes, see examples below.
}
\description{
Whether your own or someone else's writing, alex helps you find
gender favouring, polarising, race related, religion inconsiderate,
or other unequal phrasing.
}
\examples{
x <- alex(c("The boogeyman wrote all changes to the **master server**.",
             "Thus, the slaves were read-only copies of master.",
             "But not to worry, he was a cripple."))
x
x$message
}
