% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{assign_municipality_from_coordinates}
\alias{assign_municipality_from_coordinates}
\title{Assigns the municipality and region codes to points in Italy}
\usage{
assign_municipality_from_coordinates(df_sp)
}
\arguments{
\item{df_sp}{a sf data frame containing the points to label}
}
\value{
a new dataframe with the filled values
}
\description{
Assigns the municipality and region codes to points in Italy
}
