% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_utilities.R
\name{athena_createDataset}
\alias{athena_createDataset}
\title{Copies the data frame and makes it available in AWS Glue. Attention: if a table with the same name exists it will be deleted.}
\usage{
athena_createDataset(
  df,
  table_name,
  s3_base_location = "s3://rdsupport/R_datasets/",
  sep = "\\t",
  start_empty = F,
  is_spatial = F
)
}
\arguments{
\item{df}{the data frame}

\item{table_name}{the resulting table name in aws glue}

\item{s3_base_location}{the base dataset locations (default s3://rdsupport/R_datasets/)}

\item{sep}{the delimiter used (default tab)}

\item{start_empty}{remove the file leaving just the path (default F)}
}
\value{

}
\description{
Copies the data frame and makes it available in AWS Glue. Attention: if a table with the same name exists it will be deleted.
}
\examples{
athena_createDataset(df, 'my_athena_db.my_athena_table')

}
