% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_utilities.R
\name{athena_execute}
\alias{athena_execute}
\title{Executes a remote Athena command}
\usage{
athena_execute(query, verbose = F)
}
\arguments{
\item{query}{}
}
\value{
the AWS query_id
}
\description{
Executes a remote Athena command
}
\examples{

athena_execute(create_glue_table(df, foo, 's3://test/foo/'))

}
