% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_utilities.R
\name{athena_execute_fetch}
\alias{athena_execute_fetch}
\title{Executes a remote Athena query and fetch the result}
\usage{
athena_execute_fetch(query)
}
\arguments{
\item{query}{}
}
\value{
the result as a data frame
}
\description{
Executes a remote Athena query and fetch the result
}
\examples{

athena_execute(create_glue_table(df, foo, 's3://test/foo/'))

athena_execute_fetch('select * from foo')


}
