% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{beginning_and_end_of_months}
\alias{beginning_and_end_of_months}
\title{Title}
\usage{
beginning_and_end_of_months(
  initial_day = as.Date("2018-11-01"),
  end_day = Sys.Date(),
  format_string = "\%b \%Y"
)
}
\arguments{
\item{end_day}{}
}
\value{

}
\description{
Title
}
