% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{changeColName}
\alias{changeColName}
\title{Changes the name of a column in a data.frame}
\usage{
changeColName(the_dataframe, oldname, newname)
}
\arguments{
\item{the_dataframe}{dataset to change}

\item{oldname}{the old column name}

\item{newname}{the new column name}
}
\description{
Changes the name of a column in a data.frame
}
