% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkNA}
\alias{checkNA}
\title{Returns a list with the proportion of missing values per column}
\usage{
checkNA(df)
}
\arguments{
\item{df}{The input data frame}
}
\description{
Returns a list with the proportion of missing values per column
}
\examples{

hist(checkNA(iris),
  main = paste ("Histogram of the proportion of missing values"),
  xlab = "Proportion of missing values",
  ylab = "Count of variables",
  labels = TRUE)


}
