% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{ci}
\alias{ci}
\title{Confidence interval of the mean of a vector}
\usage{
ci(x, alpha = 0.05)
}
\arguments{
\item{x}{the vector}

\item{alpha}{the required two-sided confidence level (default 0.05 = 95%)}
}
\value{
the number corresponding to the one-sided confidence interval around the mean
}
\description{
At a given alpha
}
