% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_utilities.R
\name{copy_to_s3}
\alias{copy_to_s3}
\title{Copies the data frame to s3}
\usage{
copy_to_s3(df, s3_location, sep = ",")
}
\arguments{
\item{df}{the data frame}

\item{s3_location}{the s3 bucket (directories will be created)}
}
\value{

}
\description{
Copies the data frame to s3
}
\examples{
copy_to_s3(nike_locations_giulia, S3_LOCATION, sep = '\t')

}
