% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_shape_with_circles.R
\name{cover_shape_with_circles}
\alias{cover_shape_with_circles}
\title{Crea una griglia di cerchi dal raggio specificato che coprano tutta la superficie}
\usage{
cover_shape_with_circles(shape, circleRadius = 50000)
}
\arguments{
\item{circleRadius}{}
}
\value{

}
\description{
Crea una griglia di cerchi dal raggio specificato che coprano tutta la superficie
}
\examples{

circles <- cover_shape_with_circles(world.borders[world.borders$NAME=='Italy',],50000)

}
