% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{create_sql_visits_query}
\alias{create_sql_visits_query}
\title{Creates a query that returns the visits in each polygon}
\usage{
create_sql_visits_query(
  places_table_name,
  ds_from,
  ds_to,
  location_table_name = "emr.dataset_accumulo",
  places_id_column_name = "id",
  places_wkt_column_name = "wkt",
  destination_table = NULL,
  places_options = "",
  aid_options = ""
)
}
\arguments{
\item{places_table_name}{the table with the polygons}

\item{ds_from}{the day date for which we are retrieving the visits (for 1 day can be the same as ds_to)}

\item{ds_to}{the day date for which we are retrieving the visits (for 1 day can be the same as ds_from)}

\item{location_table_name}{the table that stores the ping (defaults emr.dataset_accumulo) (hardcoded columns lon, lat)}

\item{places_id_column_name}{the name of the column containing the ID of the place (defaults to id)}

\item{places_wkt_column_name}{the column containing the polygon in WKT format (defaults to wkt)}

\item{destination_table}{a destination table for the INSERT, leave default NULL if you want to fetch the rows}

\item{places_options}{can be a clause like WHERE places.bt_level4 = 'Q8'}

\item{aid_options}{can be a clause like 'AND aid in = (SELECT aid from test.audience_table)'. Should contain AND because it's placed after the select}
}
\value{
the SQL query for the visits
}
\description{
Returns visits and dwell time in each polygon. Note that the name of the IDFA column must be 'aid'
}
\examples{
create_sql_visits_query("R.mcdonalds_eevm_places", '2019-10-04', '2019-10-27')

}
