% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{cumulative_dist_plot}
\alias{cumulative_dist_plot}
\title{Cumulative distribution plot}
\usage{
cumulative_dist_plot(data_in, xlab = NULL, ylab = NULL)
}
\arguments{
\item{data_in}{Dataframe in}

\item{xlab}{Title of the X axis, default NULL}

\item{ylab}{Title of the Y axis, default NULL}
}
\description{
Cumulative distribution plot
}
\examples{

cumulative_dist_plot(iris$Sepal.Width)

}
