% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dateSinceEpoch}
\alias{dateSinceEpoch}
\title{Converts a number as Ms from epoch to POSIXct}
\usage{
dateSinceEpoch(val, tz = "CET")
}
\arguments{
\item{val}{The number of ms since Epoch}

\item{tz}{The timezone, default CET}
}
\value{
A POSIXct object with the date/time representation
}
\description{
Converts a number as Ms from epoch to POSIXct
}
\examples{

dateSinceEpoch(1231908401)

}
