% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Converts an angle in degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{rad}{The angle in radians}
}
\value{
The value of the angle in degrees
}
\description{
Converts an angle in degrees to radians
}
