% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{detach_package}
\alias{detach_package}
\title{Removes a package from memory}
\usage{
detach_package(pkg, character.only = FALSE)
}
\arguments{
\item{character.only}{}
}
\description{
Removes a package from memory
}
