% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{doPlots}
\alias{doPlots}
\title{Creates histograms for categorical variables}
\usage{
doPlots(data_in, fun, ii, ncol = 3)
}
\arguments{
\item{data_in}{The input data frame}

\item{fun}{Plotting function, i.e. plotHist for categorical variables}

\item{ii}{Indices of the columns to plot}

\item{ncol}{Number of columns (can't be 1)}
}
\description{
Creates histograms for categorical variables
}
\examples{

A <- data.frame(x = rep(c("YES","NO"), 10),
                y = rep(c('R','G','B','F'), 5),
                z = rep(c("YES","NO"), 10),
                t = rep(c('R','G','B','F'), 5))
doPlots(A, fun = plotHist, ii = 1:4, ncol = 2)

}
