% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_utilities.R
\name{emr_presto_get_db_connection}
\alias{emr_presto_get_db_connection}
\title{Returns a db connection for sending DBI queries to an EMR Presto DB instance}
\usage{
emr_presto_get_db_connection(instance_dns)
}
\arguments{
\item{instance_dns}{}
}
\value{
the DB connection object to use with the standard DBI API
}
\description{
Returns a db connection for sending DBI queries to an EMR Presto DB instance
}
\examples{

cluster_id <- emr_presto_launch_instance("Test Cluster", instance_type = "c5.4xlarge", instance_count = 4)
instance_dns <- emr_presto_get_instance_dns_name(cluster_id)
con <- emr_presto_get_db_connection(instance_dns)
res <- dbSendQuery(con, 'SELECT 1')
all_rows <- dbFetch(res, -1)
emr_presto_stop_instance(cluster_id)

}
