% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_utilities.R
\name{emr_presto_get_instance_dns_name}
\alias{emr_presto_get_instance_dns_name}
\title{Gets the presto instance DNS}
\usage{
emr_presto_get_instance_dns_name(cluster_id)
}
\arguments{
\item{cluster_id}{}
}
\value{
the dns of the instance
}
\description{
Gets the presto instance DNS
}
