% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_utilities.R
\name{emr_presto_launch_instance}
\alias{emr_presto_launch_instance}
\title{Launches an EMR Presto instanct (5 x m4.4xlarge). Synchronous and can take a while!}
\usage{
emr_presto_launch_instance(
  cluster_name = "R Cluster",
  instance_type = "c5.4xlarge",
  instance_count = 4,
  query_max_memory = "150GB",
  query_max_memory_per_node = "64GB",
  max_connections = 100,
  verbose = F,
  request_as_spot_instance = F
)
}
\arguments{
\item{cluster_name}{Cluster name defaults to R Cluster}

\item{instance_type}{Type defaults to c5.4xlarge}

\item{instance_count}{Number of slaves defaults to 4}

\item{query_max_memory}{Max memory (default '150GB')}

\item{query_max_memory_per_node}{Max memory per node (default 64GB)}

\item{max_connections}{Max connections (default 100)}

\item{verbose}{Continuously logs the cluster status while starting it}

\item{request_as_spot_instance}{Tries to reserve as spot instance}
}
\value{
the cluster_id
}
\description{
Launches an EMR Presto instanct (5 x m4.4xlarge). Synchronous and can take a while!
}
\examples{
cluster_id <- emr_presto_launch_instance("Test Cluster", instance_type = "c5.4xlarge", instance_count = 4)
instance_dns <- emr_presto_get_instance_dns_name(cluster_id)
con <- emr_presto_get_db_connection(instance_dns)
res <- dbSendQuery(con, 'SELECT 1')
all_rows <- dbFetch(res, -1)
emr_presto_stop_instance(cluster_id)

}
