% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presto_utilities.R
\name{emr_presto_stop_instance}
\alias{emr_presto_stop_instance}
\title{Stops an emr presto instance}
\usage{
emr_presto_stop_instance(cluster_id)
}
\arguments{
\item{cluster_id}{the cluster id of the instance}
}
\value{
the cluster status
}
\description{
Stops an emr presto instance
}
