% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ends_of_months}
\alias{ends_of_months}
\title{Returns a structure containing the end of each month}
\usage{
ends_of_months(
  initial_day = as.Date("2018-11-01"),
  end_day = Sys.Date(),
  format_string = "\%b \%Y"
)
}
\arguments{
\item{end_day}{}
}
\value{

}
\description{
Returns a structure containing the end of each month
}
