% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{error.bars}
\alias{error.bars}
\title{Creates a barplot with error bars}
\usage{
error.bars(yv, z, labels)
}
\arguments{
\item{yv}{the values of the bar chart}

\item{z}{the confidence levels}

\item{labels}{the labels}
}
\value{

}
\description{
Straight from the R-book 2nd edition
}
