% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expandDataFrame}
\alias{expandDataFrame}
\title{Expands a data frame by merging with a MxN permutation data.frame filled with 0 or other}
\usage{
expandDataFrame(df, namex, x, namey, y, namez, z.val = 0)
}
\arguments{
\item{df}{input data frame}

\item{namex}{variable name for rows}

\item{x}{vector of values of rows}

\item{namey}{variable name for columns}

\item{y}{vector of values of columns}

\item{namez}{name of the merged field, usually a count}

\item{z.val}{value to fill the merged column with}
}
\value{
a filled data frame
}
\description{
Expands a data frame by merging with a MxN permutation data.frame filled with 0 or other
}
\examples{
heatmap_pivot <- expandDataFrame(visit_pivot, 'hour', c(0:23), 'wday', c(1:7), 'count')
heatmap_visits <- reshape(heatmap_pivot,  idvar="hour",timevar="wday",direction="wide")
heatmap_visits$hour <- NULL

}
