% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expandDataFrame1d}
\alias{expandDataFrame1d}
\title{Expands a data frame by merging with a data.frame of length M filled with 0 or other}
\usage{
expandDataFrame1d(df, namex, x, namez, z.val = 0)
}
\arguments{
\item{df}{input data frame}

\item{namex}{variable name for rows}

\item{x}{vector of values of rows}

\item{namez}{name of the merged field, usually a count}

\item{z.val}{value to fill the merged column with}
}
\value{
a filled data frame
}
\description{
Expands a data frame by merging with a data.frame of length M filled with 0 or other
}
