% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extractWindowAroundIndices}
\alias{extractWindowAroundIndices}
\title{returns windows aroung increasing series by group indices}
\usage{
extractWindowAroundIndices(df, windowLength, startIndices, groupingVariable)
}
\arguments{
\item{df}{Input dataframe, should contain a grouping column called groupingVariable}

\item{startIndices}{Should contain the indices as start from the grouped series}

\item{groupingVariable}{The name of the column that will be used for the inned join}

\item{windowLenght}{Must be an odd number}
}
\value{
A new data frame containing just the matching columns
}
\description{
returns windows aroung increasing series by group indices
}
\examples{
A <- data.frame(name = "A", valNumeric = c(1:9), valCategorical = c("YES","NO","MAYBE"))
B <- data.frame(name = "B", valNumeric = c(5:100), valCategorical = c('YES','NO'))
C <- data.frame(name = "C", valNumeric = c(5:7), valCategorical = c('green','green','blue'))
df <- rbind(A,B,C)
startIndices <- data.frame(name = c("A","B","C"), index = c(3,10,2))
ggplot(extractWindowAroundIndices(df, 5, startIndices, "name"), aes(position, valNumeric, color = name)) + geom_line()
}
