% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extrapolate_with_confidence}
\alias{extrapolate_with_confidence}
\title{Extrapolate a number using pro}
\usage{
extrapolate_with_confidence(
  population_size,
  percentage_sampled,
  number_of_positives,
  conf_int
)
}
\arguments{
\item{population_size}{The true size of the population}

\item{percentage_sampled}{Percentage of the population sampled}

\item{number_of_positives}{How many positives have been found}

\item{conf_int}{Can be any of the classical two-sided values: 0.8, 0.9, 0.95}
}
\value{
Returns a list containing:
        - proportion of the sample that was marked positive
        - confidence interval around number_of_positives
        - confidence interfal around the true population
}
\description{
Extrapolate a number using pro
}
\examples{


}
