% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stop_locations.R
\name{find_stop_locations}
\alias{find_stop_locations}
\title{Detects the stop locations from a time series of GPS positions}
\usage{
find_stop_locations(
  sdf,
  columnsToKeep = "",
  use_centroid = T,
  max_distance_mt = 30,
  min_elements_in_cluster = 2
)
}
\arguments{
\item{sdf}{containing lat/long (crs = 4326)}

\item{columnsToKeep}{chr vector containing the columns to keep for other purposes}

\item{use_centroid}{wether to use the centroid or the last point for distance computation}

\item{max_distance_mt}{the maximum distance moved before the algorithm stops clustering}

\item{min_elements_in_cluster}{min close points that contribute to forming a cluster}
}
\value{
Returns a spatial data frame containing the summary for the clusters
}
\description{
Returns a spatial data frame containing stop locations in lat/long, min and max time,
from a spatial frame with lat/long (must have crs set to 4326, lat/long).
}
\note{
Possible extensions:
      - accuracy that gets lower when entering a building, allowing for
      clusters made by just one element
      - previous point computation based on a small movement from the centroid
      - could use speed information: whenever the speed is higher than walking speed, stop clustering
}
