% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean of a vector}
\usage{
geometric_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{the vector}

\item{na.rm}{ignores NAs}
}
\description{
Geometric mean of a vector
}
\examples{

geometric_mean(c(1,1,3,6,1,1,2,1))

}
