% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{getmode}
\alias{getmode}
\title{returns the mode of a numeric vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{Input vector}
}
\value{
The mode of the vector
}
\description{
returns the mode of a numeric vector
}
\examples{
v <- c(2,1,2,3,1,2,3,4,1,5,5,3,2,3)
result <- getmode(v)
print(result)

}
