% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_utilities.R
\name{hdfs_uploadFile}
\alias{hdfs_uploadFile}
\title{Uploads a local file to HDFS through the webhdfs API}
\usage{
hdfs_uploadFile(
  localPath,
  remotePath,
  hdfsUri = "http://192.168.0.130:9870/webhdfs/v1"
)
}
\arguments{
\item{localPath}{path to the local file}

\item{remotePath}{path to the remote file}

\item{hdfsUri}{URI of the webhdfs API, default http://192.168.0.130:9870/webhdfs/v1}
}
\value{

}
\description{
Uploads a local file to HDFS through the webhdfs API
}
