% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive_utilities.R
\name{hive_execute}
\alias{hive_execute}
\title{Executes a query on local hive}
\usage{
hive_execute(query)
}
\arguments{
\item{query}{The query}
}
\value{

}
\description{
Executes a query on local hive
}
