% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{install_library}
\alias{install_library}
\title{Installs and loads a library}
\usage{
install_library(libname)
}
\arguments{
\item{libname}{the lobrary name}
}
\description{
Installs and loads a library
}
