% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Returns the nan elements for a data frame}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{the data frame}
}
\value{
a boolean matrix
}
\description{
Returns the nan elements for a data frame
}
\examples{
df[is.nan(df)] <- 0

}
