% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_weekend}
\alias{is_weekend}
\title{returns wether the current date is a weekend day or not}
\usage{
is_weekend(ds)
}
\arguments{
\item{ds}{the input date}
}
\value{
T or F if it's weekend or not
}
\description{
returns wether the current date is a weekend day or not
}
