% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy.country}
\alias{italy.country}
\title{Shape of the Italian country}
\format{
A simple feature collection frame with 1
}
\usage{
italy.country
}
\description{
Uses the ISTAT 2018 population data
}
\keyword{datasets}
