% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy.municipalities}
\alias{italy.municipalities}
\title{Shape and population data for the municipalities in Italy}
\format{
A simple feature collection frame with 8094 rows
}
\usage{
italy.municipalities
}
\description{
Uses the ISTAT 2018 population data
}
\keyword{datasets}
