% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy.provinces}
\alias{italy.provinces}
\title{Shape and population data for the provinces in Italy}
\format{
A simple feature collection frame with 107 rows
}
\usage{
italy.provinces
}
\description{
Uses the ISTAT 2018 population data
}
\keyword{datasets}
