% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy.regions}
\alias{italy.regions}
\title{Shape and population data for the regions in Italy}
\format{
A simple feature collection frame with 20 rows
}
\usage{
italy.regions
}
\description{
Uses the ISTAT 2018 population data
}
\keyword{datasets}
