% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy.zipcodes}
\alias{italy.zipcodes}
\title{Italian zipcodes}
\format{
A simple feature collection with the polygons and the zipcodes
}
\usage{
italy.zipcodes
}
\description{
Contains the (cleaned) dataset from purchased Edilmapstore
}
\keyword{datasets}
