% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{kmToLongitudeDegrees}
\alias{kmToLongitudeDegrees}
\title{Computes how many degrees of longitude is a distance at a latitude}
\usage{
kmToLongitudeDegrees(distance_km, atLatitude = 0)
}
\arguments{
\item{distance_km}{The distance in Km}

\item{atLatitude}{The degrees of latitude}
}
\value{
The angle of longitude in degrees
}
\description{
Computes how many degrees of longitude is a distance at a latitude
}
