% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utilities.R
\name{make_italy_grid}
\alias{make_italy_grid}
\title{Creates a grid of polygons of a given size that covers the whole italian country}
\usage{
make_italy_grid(linear_meters = 10000)
}
\arguments{
\item{linear_meters}{the size of the square cell in meters (10km default)}
}
\value{
an sf object of polygons, crs 4326
}
\description{
Creates a grid of polygons of a given size that covers the whole italian country
}
