% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{marginal_histogram_plot}
\alias{marginal_histogram_plot}
\title{Marginal histogram scatterplot}
\usage{
marginal_histogram_plot(
  data_in,
  xname,
  yname,
  margin_type = "histogram",
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data_in}{Dataframe in}

\item{xname}{Name of the column for the X axis}

\item{yname}{Name of the column for the Y axis}

\item{margin_type}{Can be "histogram" "density" "violin" or "boxplot"}

\item{xlab}{Title of the X axis, default NULL}

\item{ylab}{Title of the Y axis, default NULL}
}
\description{
As in page 134 of "The Visual Display of Quantitative Information" by E.R. Tufte
}
\examples{

marginal_histogram_plot(iris, "Sepal.Width", "Petal.Length", "Sepal Length" , "Petal Length")

}
