% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_search_at_points.R
\name{nearby_search_at_points}
\alias{nearby_search_at_points}
\title{Does a find nearby places API call iterating over all the points that are in the centers_lonlat.
Pay attention because it used 3 SKU (check billing here)}
\usage{
nearby_search_at_points(
  centers_lonlat,
  api_key,
  keyword,
  radius_mt = 50000,
  wait_time = 0,
  verbose = T
)
}
\arguments{
\item{centers_lonlat}{The centers of the searches in 50 km radius}

\item{api_key}{API Key for Google Places}

\item{keyword}{Keyword, will be URL encoded by this method}

\item{radius_mt}{Radius, normally 500000}

\item{wait_time}{Eaiting time between requests, defaults to 0 seconds}

\item{verbose}{Wether to be true or not}
}
\value{
A lists of all the answers
}
\description{
Does a find nearby places API call iterating over all the points that are in the centers_lonlat.
Pay attention because it used 3 SKU (check billing here)
}
