% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{nthDecimalDigit}
\alias{nthDecimalDigit}
\title{Returns the nth decimal digit from the number x}
\usage{
nthDecimalDigit(x, position)
}
\arguments{
\item{x}{input number}

\item{position}{can be between 0 and ~ +20}
}
\value{
the number at the decimal position
}
\description{
Returns the nth decimal digit from the number x
}
\examples{

x <- 1234.1234567
floor((x * 10^5) \%\% 10)

}
