% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ alexr }
\alias{ alexr }
\alias{ alexr }
\title{ Catch Insensitive, Inconsiderate Writing }
\description{
Whether your own or someone else’s writing, alex helps you find gender favouring, polarising, race related, religion inconsiderate, or other unequal phrasing.
}
