% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{plotHist}
\alias{plotHist}
\title{Plots a histogram of column i}
\usage{
plotHist(data_in, i)
}
\arguments{
\item{data_in}{The input data frame}

\item{i}{The index of the column}
}
\description{
Plots a histogram of column i
}
\examples{

A <- data.frame(x = rep(c("YES","NO"), 10))
plotHist(A,1)

}
