% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_functions.R
\name{poissonBinomialExpectedValue}
\alias{poissonBinomialExpectedValue}
\title{Returns the expected value and confidence intervals for a poisson-binomial distribution (large number of samples)}
\usage{
poissonBinomialExpectedValue(p, n, alpha = 0.025)
}
\arguments{
\item{p}{probabilities for each group}

\item{n}{weights (number of samples for each group)}

\item{alpha}{one-sided significance level default 0.025}
}
\value{
a vector containing expected_value, expected_positives_proportion, confidence_interval_proportion
}
\description{
Returns the expected value and confidence intervals for a poisson-binomial distribution (large number of samples)
}
\examples{
probabilita_cella <- c(0.53,0.52,0.78,0.64)
campione <- c(10,3,50,15)
poissonBinomialExpectedValue(probabilita_cella, campione)

}
