% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{propByRow}
\alias{propByRow}
\title{Returns a data frame with the porportions per row}
\usage{
propByRow(df, nan.value = 0)
}
\arguments{
\item{df}{the data frame}

\item{nan.value}{the value to use for divisions by zero}
}
\value{
a data frame
}
\description{
Returns a data frame with the porportions per row
}
\examples{
df <- popByRow(df)

}
