% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Converts an angle in radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{The angle in radians}
}
\value{
The value of the angle in degrees
}
\description{
Converts an angle in radians to degrees
}
