% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{range_frame_plot}
\alias{range_frame_plot}
\title{Range-frame scatterplot}
\usage{
range_frame_plot(data_in, xname, yname, xlab = NULL, ylab = NULL)
}
\arguments{
\item{data_in}{Dataframe in}

\item{xname}{Name of the column for the X axis}

\item{yname}{Name of the column for the Y axis}

\item{xlab}{Title of the X axis, default NULL}

\item{ylab}{Title of the Y axis, default NULL}
}
\description{
As in page 131 of "The Visual Display of Quantitative Information" by E.R. Tufte
}
\examples{

range_frame_plot(iris, "Sepal.Width", "Petal.Length", "Sepal Length" , "Petal Length")

}
