% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_csv_dir}
\alias{read_csv_dir}
\title{Reads a directory of files with a given extension and creates a dataset out of it}
\usage{
read_csv_dir(
  input_path,
  input_pattern = "*.csv",
  recursive = F,
  sep = ",",
  header = T,
  add_file_column = T,
  verbose = T
)
}
\arguments{
\item{input_path}{Input direcotry}

\item{input_pattern}{Input pattern to search for files (default *.csv)}

\item{recursive}{Reads recursively}

\item{header}{Files have a header or not (default T)}

\item{add_file_column}{Adds a column with the file name}

\item{csv}{The expected separator (default ',')}
}
\value{
a new data.frame that contains all columns
}
\description{
Reads a directory of files with a given extension and creates a dataset out of it
}
\examples{
df <- read_csv_dir('/my/path/')

}
