% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_csv_dir_colnames}
\alias{read_csv_dir_colnames}
\title{Reads a directory of files with a given extension and creates a dataset out of it}
\usage{
read_csv_dir_colnames(input_path, colnames, input_pattern = "*.csv")
}
\arguments{
\item{input_path}{Input direcotry}

\item{colnames}{Names of the columns}

\item{input_pattern}{Input pattern to search for files (default *.csv)}
}
\value{
a new data.frame that contains all columns
}
\description{
Reads a directory of files with a given extension and creates a dataset out of it
}
\examples{
df <- read_csv_dir('/my/path/', colnames = c('id','value'))

}
