% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_file_as_text}
\alias{read_file_as_text}
\title{Reads a file as string}
\usage{
read_file_as_text(fileName, convertNlToSpaces = T)
}
\arguments{
\item{fileName}{the name of the file}

\item{convertNlToSpaces}{Converts newlines and tabs to spaces (useful for SQL queries)}
}
\value{
the content of the file as string
}
\description{
Reads a file as string
}
