% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{seqinclude}
\alias{seqinclude}
\title{Funcion like seq that includes the outer bound or the number after it}
\usage{
seqinclude(from, to, by)
}
\arguments{
\item{from}{Start number}

\item{to}{End number}

\item{by}{Subdivision}
}
\value{
Vector like in base::seq
}
\description{
Funcion like seq that includes the outer bound or the number after it
}
