% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain.municipalities}
\alias{spain.municipalities}
\title{Shape and population data for the municipalities in Spain}
\format{
A simple feature collection frame
}
\usage{
spain.municipalities
}
\description{
Uses the datasets at opendata.esri.es
}
\keyword{datasets}
