% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain.provinces}
\alias{spain.provinces}
\title{Shape and population data for the provinces in Spain}
\format{
A simple feature collection frame
}
\usage{
spain.provinces
}
\description{
Uses the datasets at opendata.esri.es
}
\keyword{datasets}
