% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain.regions}
\alias{spain.regions}
\title{Shape and population data for the regions in Spain}
\format{
A simple feature collection frame
}
\usage{
spain.regions
}
\description{
Uses the data at https://data.biogeo.ucdavis.edu/data/diva/adm/ESP_adm.zips
}
\keyword{datasets}
