% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{start_of_month}
\alias{start_of_month}
\title{Returns the start of the month of the date x}
\usage{
start_of_month(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Returns the start of the month of the date x
}
\examples{

end_of_last_month <- som(Sys.Date()) - 1
start_of_lasth_month <- som(som(Sys.Date()) - 1)

}
