% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{vectorToUniformDist}
\alias{vectorToUniformDist}
\title{Samples a vector from a uniform distribution, the sum adds to the mean}
\usage{
vectorToUniformDist(N, M, sd = 1)
}
\arguments{
\item{N}{the number of samples}

\item{M}{the mean of the samples}

\item{sd}{the standard deviation of the samples}
}
\value{
a vector of N numbers adding to M (can be negative)
}
\description{
Samples a vector from a uniform distribution, the sum adds to the mean
}
\examples{
vectorToUniformDist(6,10, sd = 1) / 10

}
