% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world.borders}
\alias{world.borders}
\title{Shape of the world countries}
\format{
A simple feature collection with 246 rows
}
\usage{
world.borders
}
\description{
Uses the datasets at opendata.esri.es
}
\keyword{datasets}
