% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_variants.R
\name{genotype_variants}
\alias{genotype_variants}
\title{Genotype variants in a BAM file using deepSNV::bam2R.}
\usage{
genotype_variants(variants, bam, min_bq, min_mq, mask = 0)
}
\arguments{
\item{variants}{A data frame containing variant information with columns: chr, pos, ref, alt.}

\item{bam}{Path to the BAM file to query for genotyping.}

\item{min_bq}{Minimum base quality score for reads.}

\item{min_mq}{Minimum mapping quality score for reads.}

\item{mask}{Genomic regions to mask during analysis.}
}
\value{
A tibble with columns: chr, pos, ref, alt, total_depth, alt_depth, alt_vaf.
The alt_vaf column contains the variant allele frequency calculated as alt_depth / total_depth.
}
\description{
This function genotypes a set of variants by querying a BAM file and
calculating variant allele frequencies (VAF) for each variant. It supports
SNVs, insertions, and deletions, but filters out complex variants.
}
\examples{
\dontrun{
variants <- data.frame(chr = "1", pos = 1000, ref = "A", alt = "T")
genotype_variants(variants, "sample.bam", min_bq = 30, min_mq = 30, mask = 3844)
}

}
